threading

+++ hal/include/hal/accelerometer.h
@@ -12,9 +12,9 @@
 #include <stdint.h>
 
 typedef struct {
-    int16_t x;
-    int16_t y;
-    int16_t z;
+    double x;
+    double y;
+    double z;
 } AccelerometerData;
 
 void Accelerometer_initialize(void);
